KlayGE: Samples: Separate skinning and no skinning to different fxmls.

+++ KFL/include/KFL/Math.hpp
@@ -705,9 +705,9 @@ namespace KlayGE
 
 			for (IndexIterator iter = indicesBegin; iter != indicesEnd; iter += 3)
 			{
-				uint16_t const v0Index = *(iter + 0);
-				uint16_t const v1Index = *(iter + 1);
-				uint16_t const v2Index = *(iter + 2);
+				uint32_t const v0Index = *(iter + 0);
+				uint32_t const v1Index = *(iter + 1);
+				uint32_t const v2Index = *(iter + 2);
 
 				position_type const & v0XYZ(*(xyzsBegin + v0Index));
 				position_type const & v1XYZ(*(xyzsBegin + v1Index));
@@ -788,9 +788,9 @@ namespace KlayGE
 
 			for (IndexIterator iter = indicesBegin; iter != indicesEnd; iter += 3)
 			{
-				uint16_t const v0Index = *(iter + 0);
-				uint16_t const v1Index = *(iter + 1);
-				uint16_t const v2Index = *(iter + 2);
+				uint32_t const v0Index = *(iter + 0);
+				uint32_t const v1Index = *(iter + 1);
+				uint32_t const v2Index = *(iter + 2);
 
 				position_type const & v0(*(xyzsBegin + v0Index));
 				position_type const & v1(*(xyzsBegin + v1Index));
